requirements.txt
Django==1.7
-e git+https://github.com/celiao/django-rest-authemail@53815fba2cbe7ea29d7d35719974d123acc72ae6#egg=django_rest_authemail-master
djangorestframework==3.1.0
requests==2.5.3
Pillow==2.7.0
django-push-notifications==1.2.0
